exponent to subscript mapping reversed
Decided use a shift-by-one mapping in converting
polynomial exponents to ND-array subscripts (so
for an mvpoly_dmd polynomial p, p(0,0) refers to
the constant term (and is stored in p.coef(1,1)),
p(2,1) refers to the x^2 y term (and is stored in
p.coef(3,2)), and so on.

This is the reverse to the Matlab/Octave convention
of storing a univariate polynomial with the constant
term last in the vector and the coefficeint of the
highest order monomial first (why would you do
that?).